Add tests with initial disturbances
